+++ src/main/java/firok/mtim/MoreTinkersMaterials.java
@@ -1,24 +1,38 @@
 package firok.mtim;
 
+import firok.mtim.common.Alloys;
+import firok.mtim.common.Fluids;
+import firok.mtim.common.TCMaterials;
+import firok.mtim.world.WorldGen;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.common.registry.GameRegistry;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.Logger;
 
 @Mod(
 		modid = MoreTinkersMaterials.MOD_ID,
 		name = MoreTinkersMaterials.MOD_NAME,
-		version = MoreTinkersMaterials.VERSION
+		version = MoreTinkersMaterials.VERSION,
+		dependencies = "required-after:tconstruct@[1.12.2-2.13.0.171,);required-after:mantle@[1.12-1.3.3.56,)"
 )
 public class MoreTinkersMaterials
 {
 
 	public static final String MOD_ID = "mtim";
 	public static final String MOD_NAME = "MoreTinkersMaterials";
-	public static final String VERSION = "0.1.1";
+	public static final String VERSION = "1.12.2-0.1.2";
 
 	@Mod.Instance(MOD_ID)
 	public static MoreTinkersMaterials INSTANCE;
+
+	private static Logger logger;
+	public static void log(String content)
+	{
+		logger.log(Level.INFO,content);
+	}
 
 	/**
 	 * This is the first initialization event. Register tile entities here.
@@ -27,7 +41,17 @@ public class MoreTinkersMaterials
 	@Mod.EventHandler
 	public void preinit(FMLPreInitializationEvent event)
 	{
-//		slimeknights.tconstruct.TConstruct.
+		logger = event.getModLog();
+		Fluids.register();
+
+		//
+		//  Blocks.register(false);
+		//  Blocks.registerItems();
+		//  Items.register();
+		//
+		//  proxy.initConfig();
+		//
+		TCMaterials.packMaterials();
 	}
 
 	/**
@@ -36,7 +60,24 @@ public class MoreTinkersMaterials
 	@Mod.EventHandler
 	public void init(FMLInitializationEvent event)
 	{
+		//  proxy.registerModels(); // Registers models on the client side
+		//  proxy.regsiterKeyBindings();
+		//
+		//  Fluids.registerfromItem(); // Registers some special smeltery recipes (not alloying)
+		GameRegistry.registerWorldGenerator(WorldGen.getInstance(), 100);
+		//  // GameRegistry.registerFuelHandler(new FuelHandler());  Registeres fuels' burn times
 
+		//  Blocks.register(true);
+		//
+		Alloys.registerAlloys();
+		TCMaterials.packMaterials();
+		//
+		//  SmeltingRegistry.register(); // Registers smelting recipes
+		//  CraftingRegistry.register(); // Registers crafting recipes
+		//
+		//  for (MaterialIntegration m : integrateList) {
+		//      m.integrate();
+		//  }
 	}
 
 	/**
@@ -45,6 +86,8 @@ public class MoreTinkersMaterials
 	@Mod.EventHandler
 	public void postinit(FMLPostInitializationEvent event)
 	{
-
+//		proxy.registerBookPages();
 	}
+
+
 }