+++ src/main/java/firok/mtim/util/CompoBow.java
@@ -1,0 +1,15 @@
+package firok.mtim.util;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface CompoBow
+{
+	float drawspeed();
+	float range();
+	float bonusDamage();
+}