+++ changelog.md
@@ -1,10 +1,10 @@
 # 变更日志 Changelog
 
-## 0.1.0 _2019-11-17 13:00_
+## 0.1.0 _2019-11-18 20:30_
 
 * 初始化开发环境 setup workspace
 
-## 0.1.1 _2019-11-17 14:00_
+## 0.1.1 _2019-11-17 19:30_
 
 * 增加矿物方块基类 added BlockOre class
 * 增加方块和物品管理类 added Blocks and Items class
@@ -15,3 +15,15 @@
   * 辰砂 cinnabar item
   * 恒星金属矿石 stellar ore block
   * 恒星金属锭 stellar ingot
+
+## 0.1.2 _2019-11-18 20:30_
+
+* 增加匠魂材料的注册管理器 added registry handler of TC contents
+* 增加液体 added fluids
+  * 熔融皇家合金 molten royal alloy
+* 增加方块和物品 added blocks and items
+  * 辰砂块 cinnabar block
+  * 墨粉块 ink block
+  * 恒星金属块 stellar metal block
+  * 反重力水晶 anti-grav crystal item
+* 增加世界生成器 added world generator